Revert "Create new package "extensions" and use it for an improved API"
