+++ Figaro/src/main/scala/com/cra/figaro/language/ApplyC.scala
@@ -1,13 +1,11 @@
-package com.cra.figaro.extensions
-
-import com.cra.figaro.language._
+package com.cra.figaro.language
 
 /**
   * This class is a workaround for adding easier type inference to Apply class
   * Since Apply object is already using apply method with different number of arguments
   * it is not easy to find a workaround adding currying support to apply methods in Apply class
   */
-object Apply {
+object ApplyC {
   /**
     * Application of a function to one argument.
     */
